use home-manager as system package and wrap per user
this creates a base state for user configuration on system rebuild, and
copies the configuration used into user's `$HOME` if there is none on
each call to `home-manager`.

unfortunately the user still has to call `home-manager` at least once to
get `.config` into its home directory, but this may be omitted if
copying is not done in a wrapper but a user service instead.

since the user details are private, this part is not visible on the
public branch.