fix user service to copy config
there are two major problems, one of them a show stopper:
1. `builtins.fetchGit` does not preserve `.git` (it takes the archive
   from a cloned repo). same goes for using `builtins.fetchTarball` on
   a GitHub `tar.gz`, because that simply fetches such an archive. this
   can be worked around by literally `tar`ing a repository [1] and
   hosting it somewhere, although it would be cumbersome. the other
   option is to use `pkgs.fetchgit` with overridden hashing as in [2]
   note that this will come with all the problems of
   non-deterministic `leaveDotGit` from #8567, and may produce new
   system generations even if absolutely nothing changed semantically.
2. `home-manager` system package always activates the configuration
   given, which will override changes in the user directory. as such,
   the `copy-config` service I introduced will overwrite user's config
   with what was activated. the reason I believed otherwise is probably
   that earlier I did not use `nixos-rebuild --options tarball-ttl 0`,
   and did not properly reset $HOME before testing, such that
   user-changed state persisted between rebuilds. assuming we solve 1.
   a system rebuild will force a user into the public head revision of its
   configuration. this may be tolerable, but does not satisfy the stated
   goal.

maybe a solution to both problems is replacing `home-manager.nixos` with
a custom systemd service which, instead of taking a user config
expression, takes a reference to a git repo `builtins.fetchGit` style (i.e.
optionally impure), clones it to find out where the user wants
configuration to live (assumes a known attribute to hold a path relative
to `$HOME`), and only if it is not populated by a git repo[3], copies
the repo to that location and activates that configuration.

[1]: https://github.com/NixOS/nixpkgs/issues/8567#issuecomment-133409465
[2]: http://chriswarbo.net/projects/nixos/useful_hacks.html#fetch-latest-git
[3]: or any other condition sufficiently refined but still flexible
enough to prevent existing configuration from being overwritten. maybe
checking existence of directory is enough, since the use case here is to
set up fresh systems.