add custom systemd service for user config
this is essentially following the main nixos manual advice to use systemd user
services for all post-rebuild setup, and it works.

a more sophisticated approach would be to `git clone` to a tempdir and
ask the repository where it should live within $HOME, and where
`home-manager` config should be taken from within the repository. but
I feel this would be too much configuration instead of relying convention. having
everything in default locations is not just easier to handle, but also
allows skipping `git clone` altogether if `.config` is already populated.