separate machine-specifics into module
also set up GUI and browser for initial testing.

to rebuild the system one now must run

    NIXOS_CONFIG=<repo>/machines/<machine> nixos-rebuild switch

there needs to be some more work done on bootstrapping the system with
one command. this may be in the form of a shell script living at the top
level of the repository, taking the machine ID. there is a high risk of
overthinking it because of all the underlying assumptions, so I suppose
the script should be more like a note to the user about what to do.

rebuilding also needs to be more convenient when logged in as a regular
user. this would probably be realised through machine-specific wrapper
modules on the `home-manager` configuration, giving the user a wrapper
around `nixos-rebuild` - and also around `home-manager` to take the
config file from the location where it was initially set up.