enable sound with pulseaudio
this still does not provide media keys.

unfortunately `sound.mediaKeys.enable` does not what one would expect,
and is therefore broken altogether.

while the key codes are correct, without `pulseaudio` the `amixer`
wrapper does not find a `Master` device (there are just raw devices),
and with `pulseaudio` the `actkbd` key listener fails to issue a command
because it runs as root (although `Master` would be available) and
`pulseaudio` runs as a user service expecting commands from the correct
user.

the targeted workaround here is replacing the `actkbd.nix` module (using
`disabledModules`) with a patched version that also allows binding keys
on a user-instance of `actkbd`. the only extra requirement is that the
user in question is member of group `input`.