swap alt and super keys
note that I also physically swapped the keys on the X240 to be `alt,
win`. I find it more comfortable to have GUI-related keybindings under
the left thumb, and also like the default _physical_ layout intended for
xmonad. but binding `alt` to GUI actions precludes emacs use, so the
code binding needed to be swapped with `super`.

also note that the change also must apply to TTY operation to
be consistent for all usage scenarios. nothing worse than constantly
changing keyboard layouts.