enable media keys with pulseaudio
this requires running `actkbd` per user, hence we replace the original
`actkbd.nix` with a custom version allowing for user bindings, and a
user service instance of `actkbd` using these bindings.

getting here was something like a full day of work: research, trial and
error, banging head against screen. the terror of this particular
instance stems from the fact that udev `add` events at system startup
are seemingly are not triggering user services. this requires a
workaround to collect the respective devices and instantiate
the services initially intended for them [1].

it is not clear if this really is a bug as perceived, but manually triggering
the rule works as intended. I could not find bug reports or any written
record related to the problem.

other than that, systemd configuration syntax is non-obvious, just as
interaction between udev and systemd. it did not help at all that the
`stable` channel does not employ best practice in using it (e.g.
escaping path names).

[1]: as found here <https://gist.github.com/ephemient/2f4e596fe68e1f13cbde26b8979eb058>