allow `udisks2` to mount filesystems from `systemd` unit
seems like one has to explicitly allow udisks2 to mount file systems
from other "seats", in this case `systmed`:
<https://github.com/coldfix/udiskie/issues/165#issuecomment-429568341>

this is a working but incomplete solution, as the permissions to
automatically mount a filesystem should be narrowed down to what is
absolutely required to do the job. probably this will stick for a while
nonetheless, because having a working system and more features is more
important than fixing up these issues.