use grub as bootloader, show boot splash
this does not display grub menu unless `Esc` is pressed.

there is still an issue with boot script messages leaking [1] before
`plymouth` starts.

GRUB2 is not great. it forces a welcome message upon us, which is only
effectively avoided by patching the source [2]. also the background
image is shown even if the menu is skipped, probably because it persists
in the buffer where it was prepared.

another thing is showing the boot menu only on keypress. although there
exists such a feature (or at least the possibility to make it happen),
it only works with BIOS [3]. on my X240 the BIOS introduces more delay
than is saved by avoiding the 1s timeout needed under UEFI. since the
inconvenience of pressing `Esc` vs. holding `Alt` or `Shift` is minor,
and the background flash bug exists independent of boot mode such that
there is no incentive to have eye candy in the boot menu, I stick to
UEFI where possible.

systemd-boot has an annoying UI, and although it supports a high
resolution mode, there is still a screen flicker (even with
i915.fastboot=1).

[1]: https://github.com/NixOS/nixpkgs/issues/32555
[2]: http://rabexc.org/posts/grub-shush
[3]: https://github.com/NixOS/nixpkgs/issues/2776