pin nixpkgs declaratively
this is more of a draft, but it works for an existing system. see TODOs
to generalise this for installation from bootable image.

refactor the repository to make a clear distinction between NixOS
modules (which provide configuration *options*) and actual configuration
*definitions* for `machines`. the ones shared by multiple machines now
live in `common`.

update TODOs based on new insights.