copy system repository to /etc/nixos when bootstrapping
I struggled with centralising the repository for sudoers, as it as has
a number of disadvantages, such as the potential need to set up the whole
git/ssh/editor stuff for `root`. there is not much wrong with having the repo
under a regular user's control, since one needs `sudo` to switch to
a configuration anyway. also it is much more convenient to edit.
the other major problem is that building the initial configuration at
bootstrap time from anywhere arbitrary will point `nixos-rebuild` to some
configuration that won't exist any more after the first boot.
alternatives considered:
1. use `git-crypt` only for sensitive files.
   the principle of encrypting only some data appeals to me. it would also
   solve all of the automation problems, since pulling requires no secrets.
   also much of the work can be documented publicly for better knowledge
   exchange, which would be lost on a completely private repository. the
   problem with it is that the system is so brittle and error prone.
2. let install script copy the repo based on configuration provided.
   just moving it to some user's `$HOME` does not seem wise, it should
   be done as part of some automatic environment setup under `nix` control.
3. clone via a user service.
   in the current setup this would require a rebuild, such that the wrapper
   looks for input at the right location. it must happen after login, such
   that `sudo` password can be prompted. not sure if this would really happen
   or if a user service even works if it runs `sudo` commands.
4. build more advanced wrapper for `nixos-rebuild`.
   `builtins.fetchGit` from repo HEAD by default (assuming
   `git-remote-gcrypt` is used). take alternative branch name or local repo
   as parameter.
   wrapper is still tied to the repo structure and needs careful (manual)
   rebuild if that changes - as is already the case.
   the whole thing only works if `git-remote-gcrypt` manages to call the
   current user's `gpg-agent` when called with sudo (I suspect it does).
   probably will not work with `autoUpgrade`, as that is run by `root`, so it
   won't have `gpg-agent` available.
5. create `nixos-rebuild` wrapper in user config.
   any sudoer can instantiate arbitrary system configuration anyway, so why
   try to centralise some presumed default behavior? we have to assume
   responsible behavior from admins. in this setup a user service clones the
   system repo and I think it should properly prompt for GPG key password.
   the wrapper is build on knowledge about the repo structure, but is
   otherwise independent. no system rebuild is required if the repo changes
   for some reason. `autoUpgrade` will not work here.

on the other hand one could copy the repo to a well-known location (e.g.
/etc/nixos) and prepare its permissions such that group `wheel` can edit it
from the given user's environment as needed. this has to be done by the
installation script, as that is the only attended portion before user login.
the advantage here is that we then can have `autoUpgrade. somewhat secondary
we do not have to clone the repo twice.

now the bootstrap sequence is supposed to look roughly like this:

- boot NixOS from USB
- mount USB stick with secrets
- run `$secrets/install` (TODO)
  - clones encrypted system repository
  - asks for password to GPG key on USB stick
  - runs `nixos/install`
- set `root` password
- reboot
- system service sets up user environment before login
  - clones *public* user repo, this is unattended
  - imports GPG key to password store
- user service clones password repo after login (TODO)
  - asks for password to imported GPG key

this way the only dependency is the USB stick with a GPG key to the repository
containing system configurations. everything else is derived from there.

note that the user repo must be public in this setup. for me this is not an
issue as my user environment does not contain sensitive data that is not
already public anyway. if you want an encrypted user repo it would have to be
pre-cloned/decrypted during the attended bootstrap phase and moved to
a location from which it can be moved by the system service setting up the user
environment.

follow-up idea: keep a copy of each repository together with the secrets. then
the respective installation steps will `git pull --force` on them if internet
connection is available, otherwise use whatever is already there. if we could
also store the system and user environment closures after each successful
build, and point `nix` to that local store on installation, that USB stick in
conjunction with a bootable NixOS image would be entirely self-sustainable.
even more convenient would be to keep the mutable data on the boot image.