set environment variables for `gpg` and `pass`
I do not like this approach at all, because it pollutes the environment.
none of these settings should be session variables, but right now
this is the simplest approach to write and read. it would be really
cumbersome to pick out each relevant binary using these values and
wrap them appropriately. even if there are just a few I am aware of, it
would be hard to avoid breakage due to omissions and annoying to follow
through for all the programs having this issue.