rewrite install script in bash
the script did not work on macOS because `fish` prepends the contents of
`/etc/paths` and `/etc/paths.d/*` to `$PATH` [1], resulting in BSD
`readlink` to be used instead of the one from `nixpkgs`. this is fixed
on the master branch [2], but the patch is staged for 3.1.0 (!), while
the most recent release is 3.0.2.

well, also the script was badly broken, I never ran it since changing it
the last time...

[1]: https://github.com/fish-shell/fish-shell/issues/5456
[2]: https://github.com/fish-shell/fish-shell/pull/5637