notify on critical battery
it is not yet as nice as it could be. we cannot call arbitrary commands
in `xmobarrc` or `xmonad.hs` for example. for that we'd need to replace
the command names with `/nix/store/...` and that would mean some
processing of the raw source, but more importantly some probably illegal
syntax for marking up those references, which would mess up syntax
checking/highlighting in the respective files.

specifically I would like to display the remaining battery time in
a notification, but I have to call `acpi` for that, which would require
the above procedure.