reorganize modules into machine profiles
this is the preliminary step for defining a user environment on my home
server `una`. since graphical software is not needed there, it has to
be unbundled from the generic linux configuration.

note that I took the "workstation" nomenclature from [1]. I do not
really like it, as it is not perfectly precise, but it appears to be the
clearest sounding name for a machine with human interface devices.
alternatives considered: client, terminal, console, frontend

client:
  even if I will have servers at some point, these machines are not
  necessarily clients to anything specifically.

terminal:
  while this is technically meaningful to use here in the sense of
  "endpoint", the association with "text-only user interface" is too
  strong for my taste. but thinking in these terms also reveals
  a potential weakness in the current architecture: why should
  a graphical configuration *include* everything text-based? it only
  does so because this is how I use my systems. the point of importing
  some configurations into others was to have a consise machine
  specification. the directory stucture is supposed to (partially)
  reflect the these inclusion dependencies, making the import line
  somewhat self-explanatory in that regard, but of course this is
  arbitrary. it could just as well be `modules/linux/graphical/` instead
  of `modules/linux/workstation/graphical`. one could eliminate the
  import of more common parts into the more specific ones altogether,
  and instead be more precise about required features in the machine
  definitions.

console:
  same as "terminal", but additionally here is a strong association with
  a clunky physical console as a HID, which is slightly more sensible with
  respect to how the configuration is structured currently. I think
  using it would obscure more than it would reveal.

frontend:
  has the same problem as "client", since these machines are not front
  ends for anything specific.

I still need to strike the right balance between short machine
specification and sufficiently independent configuration modules. also
the distinction between TUI and GUI could have more sophisticated
naming. `tui`/`gui` and `text`/`graphical` are not self-explanatory,
`text-ui`/`graphical-ui` may be candidates.

[1]: https://github.com/dustinlacewell/dotfiles/