home-config: do not bind init to check, only require
the previous state seems to be a bug, which resulted from an untested
but otherwise meaningful change. unfortunately the whole setup was
broken, since on condition success the check service stopped - but since
the initialization was bound to it, that would not start subsequently.
the other way to fix this would be to let the check `RemainAfterExit`.
I am not aware of any meaningful tradeoff here. I can imagine negligible
differences in performance or memory consumption within `systemd`, but
the only thing I can actually measure is the number of characters used
for the service definition.