fix home-config service once again
this time I think I finally got it. at first boot the check should
succeed but remain after exit such that the init service can pick it up.
the init service binds to the check, which requires the persistence on
check succes, but also requires it such that the conditional execution
actually prevents the init from starting when the condition is not
satisfied.